Fix custom error handler
